+++ test/unit/PropertiesPicker.js
@@ -63,4 +63,70 @@ describe('PropertiesPicker', () => {
 			assert.strictEqual(getterCallsCount, 3);
 		});
 	});
+
+	describe('mergeProperties', () => {
+		it('basic', () => {
+			const picker = new PropertiesPicker(new Map([
+				[['a'], ['topScalar']],
+				[['b'], ['topNull']],
+				[['c'], ['topUndefined']],
+				[['d'], ['topUndefined']],
+				[['f'], () => {
+					return 'hi';
+				}]
+			]));
+
+			const dest = {
+				e: 'hello'
+			};
+
+			const o = picker.mergeProperties(object, dest);
+
+			assert.strictEqual(o, dest);
+			assert.strictEqual(o.a, object.topScalar);
+			assert.strictEqual(o.b, object.topNull);
+			assert.strictEqual(o.c, object.topUndefined);
+			assert.strictEqual(o.d, undefined);
+			assert.strictEqual(o.e, 'hello');
+			assert.strictEqual(o.f, 'hi');
+		});
+
+		it('replace', () => {
+			const picker = new PropertiesPicker(new Map([
+				[['a'], ['topScalar']],
+				[['b', 'c'], ['topNull']]
+			]));
+
+			const dest = {
+				a: 'hello',
+				b: {
+					c: 'world',
+					d: 1
+				}
+			};
+
+			const o = picker.mergeProperties(object, dest);
+
+			assert.strictEqual(o, dest);
+
+			assert.strictEqual(o.a, object.topScalar);
+			assert.strictEqual(o.b.c, object.topNull);
+			assert.strictEqual(o.b.d, 1);
+		});
+
+		it('deep non-object', () => {
+			const picker = new PropertiesPicker(new Map([
+				[['a', 'b'], ['topScalar']]
+			]));
+
+			const dest = {
+				a: 'hello'
+			};
+
+			const o = picker.mergeProperties(object, dest);
+
+			assert.strictEqual(o.a, 'hello');
+			assert.strictEqual(o.a.b, undefined);
+		});
+	});
 });