+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Support for `DELETE WHERE` queries
 * Support for `INSERT {object}[,{object}]` queries. Rows just was aded to end of the input stream
 * Added support for E-notation for floating numbers
+* Support for `SELECT *, [fields...]`
 
 ## `v1.3.0` (2016-12-05)